pyproject.toml
[tool.poetry]
name = "so_extreme"
version = "0.1.0"
description = ""
authors = ["aetienne-web-atrio <adrien.etienne@web-atrio.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
redis = "^4.3.4"
fastapi = "^0.85.0"
SQLAlchemy = "^1.4.41"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
psycopg2 = "^2.9.4"
asyncio = "^3.4.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
